SPEEDUP: positiveFusedLasso() introduced lots of overhead for forked processes.
RATIONALE FOR WHY THIS COMMIT MAKES THINGS MUCH FASTER:

First, mclapply() forks the current R process.  Forking is performed by the OS.
Forking basically means that each mclapply() iteration gets an exact copy of
the main R process and performs its calculations.

Next, look at get.Z() that mclapply() calls.  The get.Z() function calls
Matrix::bandSparse() and glmnet::glmnet() and a few other things.  Now,
neither 'Matrix' nor 'glmnet' are loaded in the main R process (at least
not in a fresh R 'c3co' session.  This means that they are also not loaded
in the forked processes.  So, each fork needs to load those packages.  My
benchmarking shows that this takes ~0.5-1 sec to do.

Continuing, mclapply() is called in every while iteration, meaning this
is done over and over until the algorithm converges.

This is why options(mc.cores = 1) is actually faster, because it causes
mclapply() to fall back to a regular lapply() in the current R process
where those packages will be loaded in all of the following iterations.

There is one more quirk that also slows things down when using forks;
The kronecker() call done by get.Z() triggers a complicated S4 method
dispatch lookup, which is why we see:

Note: method with signature 'dsparseMatrix#dsparseMatrix' chosen for
function 'kronecker', target signature 'dgCMatrix#dgTMatrix'.
 "sparseMatrix#TsparseMatrix" would also be valid

Now, this lookup is also costly and to minimize the overhead the 'methods'
package caches this lookup.  However, when you fork, this cache is not
yet updated, and the lookup takes place over and over in each fork.

(After this commit, I'll do another one that shows how to "fix" this too,
but I thought it would be more clear I leave it out here.)